requirements.txt
Django==4.0.6
Pillow==9.2.0
python-dotenv==0.20.0
requests==2.28.1